refactor: Supabase 서버 클라이언트 진입점을 모으고 인증 게이트를 features로 옮긴다
createSupabaseRequestClient가 cookies()를 안에서 받아 호출자는 Next의 쿠키
저장소를 몰라도 된다. 두 팩토리의 같은 requireEnv는 readSupabaseEnv 하나로
합쳤고, src/app/auth-gate.ts의 조립 코드는 src/features/auth/read-auth-gate.ts로
옮겨 클라이언트를 인자로 받는다. scripts 두 곳에 복제돼 있던 SUBSECTION 정규식도
tokens-md.mts가 export하는 하나가 됐다.
