+++ src/app/auth/logout/route.ts
@@ -1,9 +1,8 @@
-import { cookies } from "next/headers";
 import { NextResponse, type NextRequest } from "next/server";
-import { createSupabaseServerClient } from "@/shared/lib/create-supabase-server-client";
+import { createSupabaseRequestClient } from "@/shared/lib/create-supabase-request-client";
 
 export async function POST(request: NextRequest) {
-  const client = createSupabaseServerClient(await cookies());
+  const client = await createSupabaseRequestClient();
 
   await client.auth.signOut();
