+++ src/shared/lib/create-supabase-request-client.ts
@@ -1,0 +1,8 @@
+import { cookies } from "next/headers";
+import { createSupabaseServerClient } from "@/shared/lib/create-supabase-server-client";
+
+export async function createSupabaseRequestClient() {
+  const cookieStore = await cookies();
+
+  return createSupabaseServerClient(cookieStore);
+}
