+++ src/shared/lib/read-supabase-env.ts
@@ -1,0 +1,27 @@
+export type SupabaseEnv = {
+  url: string;
+  anonKey: string;
+};
+
+function requireEnv(name: string, value: string | undefined): string {
+  if (!value) {
+    throw new Error(
+      `환경 변수 ${name} 값이 비어 있다. NEXT_PUBLIC_* 값은 빌드 시점에 번들에 박히니, 배포 환경에 값을 채우고 다시 빌드해야 한다.`,
+    );
+  }
+
+  return value;
+}
+
+export function readSupabaseEnv(): SupabaseEnv {
+  return {
+    url: requireEnv(
+      "NEXT_PUBLIC_SUPABASE_URL",
+      process.env.NEXT_PUBLIC_SUPABASE_URL,
+    ),
+    anonKey: requireEnv(
+      "NEXT_PUBLIC_SUPABASE_ANON_KEY",
+      process.env.NEXT_PUBLIC_SUPABASE_ANON_KEY,
+    ),
+  };
+}
